requirements.txt
youtube-transcript-api
langchain-community
langchain-google-genai
faiss-cpu
tiktoken
python-dotenv
langchain_text_splitters